Moving itl80211_security description into a separate file
